Added DefaultSpectatorTeamID, and added replication for certain necessary variables. Also moved LobbyGameMode's post login super call to the end, as we want the blueprint stuff to be called last

+++ Source/June13/Lobby/LobbyPlayerStateInterface.h
@@ -17,5 +17,5 @@ class ILobbyPlayerStateInterface
 	GENERATED_IINTERFACE_BODY()
 
 public:
-	virtual int8 GetTeamID() = 0; //UE4 code standard advises against using just (int). GameModes should determine what each number represents
+	virtual int32 GetTeamID() = 0; //UE4 code standard advises against using just (int). GameModes should determine what each number represents
 };
